Fix double base url
The redirect plugin combines the site url and site baseurl. Because we
are using the full https for the base url as well as the url (due to
https errors when serving css and js content) the url appears twice,
this customises the redirect page so that it only ever redirects to the
home page.